separated querystringable and center into their own files
